- address review comments: re-organize code to avoid cyclic imports
