+++ sdk/turing/batch/job.py
@@ -91,10 +91,9 @@ class EnsemblingJob(ApiObject):
         """
         Fetches latest updates of this ensembling job
         """
-        from turing.session import active_session
         self.__dict__.update(
             EnsemblingJob.from_open_api(
-                active_session.get_ensembling_job(job_id=self.id)
+                turing.active_session.get_ensembling_job(job_id=self.id)
             ).__dict__
         )
 
@@ -102,9 +101,7 @@ class EnsemblingJob(ApiObject):
         """
         Terminates this ensembling job
         """
-        from turing.session import active_session
-        active_session.terminate_ensembling_job(job_id=self.id)
-
+        turing.active_session.terminate_ensembling_job(job_id=self.id)
         self.refresh()
 
     @classmethod
@@ -112,10 +109,8 @@ class EnsemblingJob(ApiObject):
         """
         Fetch ensembling job by its ID
         """
-        from turing.session import active_session
         return EnsemblingJob.from_open_api(
-            active_session.get_ensembling_job(job_id=job_id)
-        )
+            turing.active_session.get_ensembling_job(job_id=job_id))
 
     @classmethod
     def submit(
@@ -129,8 +124,6 @@ class EnsemblingJob(ApiObject):
         :param config: configuration of ensembling job
         :return: instance of ensembling job
         """
-        from turing.session import active_session
-
         job_config = turing.generated.models.EnsemblerConfig(
             version=EnsemblingJob._VERSION,
             kind=turing.generated.models.EnsemblerConfigKind(EnsemblingJob._KIND),
@@ -144,8 +137,7 @@ class EnsemblingJob(ApiObject):
         )
 
         return EnsemblingJob.from_open_api(
-            active_session.submit_ensembling_job(job=job)
-        )
+            turing.active_session.submit_ensembling_job(job=job))
 
     @classmethod
     def list(
@@ -162,14 +154,11 @@ class EnsemblingJob(ApiObject):
 
         :return: list of ensembling jobs
         """
-
-        from turing.session import active_session
-
         mapped_statuses = None
         if status:
             mapped_statuses = [turing.generated.models.EnsemblerJobStatus(s.value) for s in status]
 
-        response = active_session.list_ensembling_jobs(
+        response = turing.active_session.list_ensembling_jobs(
             status=mapped_statuses,
             page=page,
             page_size=page_size
