+++ sdk/turing/session.py
@@ -1,7 +1,6 @@
 import os
 import mlflow
 from typing import List, Optional
-
 from turing.ensembler import EnsemblerType
 from turing.generated import ApiClient, Configuration
 from turing.generated.apis import EnsemblerApi, EnsemblingJobApi, ProjectApi
@@ -197,30 +196,3 @@ class TuringSession:
             project_id=self.active_project.id,
             ensembling_job=job
         )
-
-
-active_session: TuringSession = TuringSession(
-    host="http://localhost:8080",
-    use_google_oauth=False
-)
-
-
-def set_url(url: str, use_google_oauth: bool = True):
-    """
-    Set Turing API URL
-
-    :param url: Turing API URL
-    :param use_google_oauth: whether use google auth or not
-    """
-
-    global active_session
-    active_session = TuringSession(host=url, use_google_oauth=use_google_oauth)
-
-
-def set_project(project_name: str):
-    """
-    Set active project
-
-    :param project_name: project name
-    """
-    active_session.set_project(project_name)
